+++ comp2013/src/main/java/comp2013/Controller/IController.java
@@ -68,4 +68,5 @@ public interface IController {
      * Restarts the game to its original state.
      */
     void restartGame();
+
 }