+++ comp2013/src/main/java/comp2013/View/SnakeView.java
@@ -41,7 +41,8 @@ public class SnakeView extends Application implements IView {
     private Image M_SnakeHeadImg, M_SnakeBodyImg, M_BackgroundImage;
     private SnakeFood M_SnakeFood;
     private Label M_ScoreLabel, M_CountDownLabel,
-            M_GameOverLabel, M_DefaultLabel;
+            M_GameOverLabel, M_DefaultLabel,
+            M_PausedLabel, M_PauseVolumeLabel;
     private Timeline M_Timeline;
     private Button M_RestartButton, M_MenuReturnButton,
             M_EnterNameButton;
@@ -49,10 +50,11 @@ public class SnakeView extends Application implements IView {
     private static SnakeView m_Instance;
     private double M_MusicVolume = 0.2;
     private StackPane M_DefaultPane;
+    private VBox M_DarkStripVbox;
     private TextField M_EnterNameField;
     private Timeline M_CountDownTimeline;
-
-    private boolean M_FirstEntry = true;
+    private Slider M_PauseVolumeSlider;
+    private boolean M_FirstEntry = true, M_GamePaused = false;
 
 
     public SnakeView() {
@@ -401,6 +403,75 @@ public class SnakeView extends Application implements IView {
         M_CountDownTimeline.setCycleCount(11);
         M_CountDownTimeline.play();
     }
+    @Override
+    public void pauseGame() {
+        if (!M_GamePaused) {
+            // Set the game to paused
+            M_GamePaused = true;
+            // Stop the timeline so the snake no longer moves.
+            M_Timeline.stop();
+            // Create a new vbox that acts as a transparent strip.
+            M_DarkStripVbox = new VBox(10);
+            M_DarkStripVbox.setStyle
+                    ("-fx-background-color: rgba(0, 0, 0, 0.6);");
+            // Set its size
+            M_DarkStripVbox.setMinSize((double) m_Controller.m_Model.getWidth() / 2,
+                    m_Controller.m_Model.getHeight());
+            M_DarkStripVbox.setMaxSize((double) m_Controller.m_Model.getWidth() / 2,
+                    m_Controller.m_Model.getHeight());
+
+            // Create a label for the pause screen
+            M_PausedLabel = new Label("Paused");;
+            // Add styling
+            M_PausedLabel.getStyleClass().add("label-with-padding");
+            M_PausedLabel.setStyle("-fx-text-fill: white;" +
+                    " -fx-font-size: 38; -fx-underline: true");
+
+            M_PauseVolumeLabel = new Label(String.format("Volume: %.0f%%", M_MusicVolume * 100));
+            M_PauseVolumeSlider = new Slider(0, 100,
+                    this.M_MusicVolume * 100); // min, max, initial value
+            M_PauseVolumeSlider.setShowTickMarks(true);
+
+            // Add a listener to respond to changes in the slider value and
+            // update the volume.
+            M_PauseVolumeSlider.valueProperty().addListener
+                    ((observable, oldValue, newValue) -> {
+                        M_PauseVolumeLabel.setText(String.format
+                                ("Volume: %.0f%%", newValue));
+                        // Update the local variable so that all music is synced
+                        this.M_MusicVolume = (double) newValue / 100;
+                        // Update the music object itself
+                        m_SnakeMusic.setVolume(((double)newValue / 100));
+
+                    });
+
+            // Set the size of the slider.
+            M_PauseVolumeSlider.setMinHeight((int)(m_Controller.m_Model.getHeight() / 20));
+            M_PauseVolumeSlider.setMinWidth((int)(m_Controller.m_Model.getWidth() / 3));
+
+            M_PauseVolumeSlider.setMaxHeight((int)(m_Controller.m_Model.getHeight() / 20));
+            M_PauseVolumeSlider.setMaxWidth((int)(m_Controller.m_Model.getWidth() / 3));
+            // Add styling
+            M_PauseVolumeLabel.getStyleClass().add("label-with-padding");
+            // Set the text to be white.
+            M_PauseVolumeLabel.setStyle("-fx-text-fill: white;");
+
+            M_SnakePane.getChildren().addAll(M_DarkStripVbox, M_PausedLabel,
+                    M_PauseVolumeLabel, M_PauseVolumeSlider);
+            // Set the position
+            M_PausedLabel.setTranslateY(-200);
+            M_PauseVolumeLabel.setTranslateY(-150);
+            M_PauseVolumeSlider.setTranslateY(-100);
+        }
+        else{
+            M_GamePaused = false;
+            M_SnakePane.getChildren().removeAll(M_DarkStripVbox, M_PausedLabel,
+                    M_PauseVolumeLabel, M_PauseVolumeSlider);
+            M_Timeline.play();
+        }
+    }
+
+
     @Override
     public void restartGame(){
         // Reset the rate of the timeline.